Bridge sherpa and astropy fitting
## Impact on sherpa and astropy

I think that this project would be a
very valuable addition to astropy and it would greatly increase the number of users for the Sherpa
fitting code (it would also significantly increase SAOs contribution to a package that's sometimes seen as STScI's brain child, although that's not quite right, of course). 

##  Open question: Licensing
I don't know if the Sherpa and astropy licenses are compatible and some Sherpa code could included in astropy or if this is better handled by writing an "adapter class", but that are implementation details to be worked out later.

Sherpa is still GPL and there is always a question even for optional dependencies.  There might be some wiggle room if the optional dependency is not substantially required for the main functionality but there are several other ways to deal with that. E.g. this could be a separate, new, GPL licensed, package "astropysherpabridge" with "from astropysherpabridge.optiizers import ...". All the objects would have the astropy interface, and the package could depend on both, or the adapter could be part of Sherpa.
Alternatively, the Sherpa team might consider relicense some of their code (not sure how much of the fitting is GPL OTS).
These details need to be worked out with the student as part of the project proposal.

@DougBurke : I've listed you here because this project requires a (co-) mentor from the Sherpa team I think. Would you be willing to mentor this together with e.g. @taldcroft or me?