Remove the manual edition of json files for projects
I've moved the things a bit. Now we have a collection of projects, under
which we can create different years and organisation directories to
store the different ideas of each organisation.

The directory that includes all the magic for the displays thanks to the
coala project is now under gsoc/display (it was under
gsoc/projects2017). Making it now possible to display projects in the
future years.
To be able to do so the `projects.js` file is now generated
automatically when using the `_layouts/projects`.
(At the moment generates a lot of white space within the json).

The path of the files to where the coala display looks has also been
edited, so it's taken directly from the header of each file, and as
jekyll converts it to html, the markdown-js for that bit is not needed.