Link to discussion page added for Gnuastro's projects
Until now Sachin had only put the link containing a checklist of steps
to do on the Matrix chat page. However, with newer chats coming down,
they were not easily findable.

With this commit, a sentence has been added at the end of the two
ideas the point interested students directly to the proper place to
get started.