GSOC: Add more detail on "open a PR" requirement
This change of text is based on feedback from our GSOC student this year. To them, the "need to open at least on PR" seemed burdensome. I think we agree now that this requirement is good for all (students and mentors) because it puts all students on equal footing - applications are judged based on real PRs, not the perceived quality of one school over another. However, it helps to have a short explanation in the text and it also helps to make clear that we don't want to create extra work. If an applicant has contributed to OA projects in the past, that should be sufficient and not require a new PR in the exact timeframe of the GSOC application.
@aaryapatil @suyog7130